Refactored into client
