+++ src/py_incharge/__init__.py
@@ -4,10 +4,10 @@ PyInCharge
 Unofficial python library for remotely controlling your VattenFall InCharge charging stations
 """
 
-__version__ = "0.1.0"
+__version__ = "0.1.3"
 __author__ = "Bram"
 __email__ = "b.dewit@applyai.nl"
 
-from .charger import send_remote_start
+from .client import InCharge
 
-__all__ = ["send_remote_start"]
+__all__ = ["InCharge"]
